Fix bugs in write callback for ProxySQL clients.
When update message buffers are too big writes to ProxySQL
clients fail, as binlog reader tries to perform these in a
single write() call. This behavior is particularly noticeable
when -t is enabled, and causes ProxySQL to disconnect, and
immediately connect back.

This commit reworks writes to batch in chunks of a maximum
configurable size - currently 4kb.